refactor: reuse asset name resolution across stores
